Course2_CloudDataWarehouse/CloudDataWarehous/requirements.txt
boto3>=1.28.0
psycopg2-binary>=2.9.5
configparser>=5.3.0
